Chore: create `utils/` directory and move `ScrollToTop.jsx` into it
