+++ src/main.jsx
@@ -1,6 +1,6 @@
 import ReactDOM from "react-dom/client";
 import { BrowserRouter } from "react-router-dom";
-import ScrollToTop from "./ScrollToTop";
+import ScrollToTop from "/src/utils/ScrollToTop";
 import App from "./App";
 import "./scss/style.scss";
